# doc/api.yaml
openapi: 3.0.3
...
info:
  title: WASAPhotoAPI
  version: 1.0.0
  description: |-
    API for the WASAPhoto application. <br>
    The API follows the CRUD philosophy, that maps HTTP methods TK
  contact:
    name: Emanuele Scaccia
    email: scaccia.2006749@studenti.uniroma1.it
    url: https://github.com/dev_excale/WASAPhoto
servers:
  - url: /
...
paths:
...
  /users/{user_uuid}/feed/{post_uuid}/likes:
    get:
      operationId: getLikes
      summary: Get users who liked a post
      description: Returns the list of users who liked the given post.
      tags: [ 'Like' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '200': # OK
          $ref: '#/components/responses/200-UserList'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
    put:
      operationId: likePhoto
      summary: Like a post
      description: Adds a like to the given post.
      tags: [ 'Like' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '204': # No Content
          description: Like added correctly
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
    delete:
      operationId: unlikePhoto
      summary: Remove the like from a post
      description: Removes the user's like from the given post.
      tags: [ 'Like' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '204': # No Content
          description: Like removed correctly
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
  /users/{user_uuid}/feed/{post_uuid}/comments:
    get:
      operationId: getComments
      summary: Get a post's comments
      description: Returns the list of comments made under the given post.
      tags: [ 'Comment' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '200': # OK
          $ref: '#/components/responses/200-CommentList'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
    put:
      operationId: commentPhoto
      summary: Comment a post
      description: Adds a comment to the given post.
      tags: [ 'Comment' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
      responses:
        '201': # Created
          $ref: '#/components/responses/201-Comment'
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-PostNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
  /users/{user_uuid}/feed/{post_uuid}/comments/{comment_uuid}:
    delete:
      operationId: uncommentPhoto
      summary: Remove a comment
      description: Removes the given comment from the given post.
      tags: [ 'Comment' ]
      security:
        - Bearer: [ ]
      parameters:
        - $ref: '#/components/parameters/pathUserUUID'
        - $ref: '#/components/parameters/pathPostUUID'
        - $ref: '#/components/parameters/pathCommentUUID'
      responses:
        '204': # No Content
          description: Comment removed correctly
        '400': # Bad Request
          $ref: '#/components/responses/400-BadRequest'
        '401': # Unauthorized
          $ref: '#/components/responses/401-Unauthorized'
        '403': # Forbidden
          $ref: '#/components/responses/403-Banned'
        '404': # Not Found
          $ref: '#/components/responses/404-CommentNotFound'
        '500': # Internal Error
          $ref: '#/components/responses/500-InternalError'
...
components:
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer

  ##########################
  ## Parameter components ##
  ##########################
  parameters:
    pathUserUUID:
      name: user_uuid
      in: path
      required: true
      description: UUID of the target user
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      schema:
        $ref: '#/components/schemas/UUID'
    pathPostUUID:
      name: post_uuid
      in: path
      required: true
      description: UUID of the target post
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      schema:
        $ref: '#/components/schemas/UUID'
    pathCommentUUID:
      name: comment_uuid
      in: path
      required: true
      description: UUID of the target comment
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      schema:
        $ref: '#/components/schemas/UUID'

  #######################
  ## Schema components ##
  #######################
  schemas:
    UUID:
      type: string
      description: A UUID
      example: 9a6a4415-9a50-4065-b345-8c22e8bb8c5f
      format: uuid
    Username:
      type: string
      description: A user's username
      example: excale
      pattern: '[\w\-.]+'
      minLength: 3
      maxLength: 20
    User:
      type: object
      description: Information about a user
      properties:
        username:
          $ref: '#/components/schemas/Username'
        display_name:
          type: string
          description: The user's display name.
          example: Emanuele Scaccia
        url_profile_picture:
          type: string
          nullable: true
          description: |-
            URL of the user's profile picture.
            Note: all photos are encoded with webp format.
            Generally, the url will be of the form
            `https://api.example.com/v1/users/{username}/assets/{photo_id}.webp`
          example: https://api.example.com/v1/users/excale/assets/3e270f31-6eaa-4418-ab97-7ed22f51d446.webp
        num_posts:
          type: number
          description: The number of posts the user has
          minimum: 0
          example: 4
        num_followers:
          type: number
          description: The number of followers the user has
          minimum: 0
          example: 123
        num_following:
          type: number
          description: The number of users the user is following
          minimum: 0
          example: 1234
    UserList:
      type: array
      description: A list of users
      items:
        $ref: '#/components/schemas/User'
...
    Comment:
      type: object
      description: Information about a comment
      properties:
        uuid:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
          description: The username of the user who added the comment.
          example: excale
        text:
          type: string
          description: The text of the comment.
          example: Nice photo!
        timestamp:
          type: string
          description: The timestamp of the comment, in `YYYY-MM-DD hh:mm:ss` format.
          example: 2021-01-01T00:00:00.000Z
    CommentList:
      type: array
      description: A list of comments
      items:
        $ref: '#/components/schemas/Comment'

  #########################
  ## Response components ##
  #########################
  responses:
...
    200-UserList:
      description: List retrieved correctly.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserList'
...
    200-CommentList:
      description: List retrieved correctly.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CommentList'
...
    201-Comment:
      description: Comment created correctly.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Comment'
...
    400-BadRequest:
      description: The request was invalid or malformed (e.g. missing parameter).
    401-Unauthorized:
      description: The user is not logged in.
...
    403-Banned:
      description: The user has been banned by the given user.
...
    404-PostNotFound:
      description: The specified post doesn't exist.
    404-CommentNotFound:
      description: The specified comment doesn't exist.
...
    500-InternalError:
      description: An internal error occurred.